fix: move files to LOCALAPPDATA if it is installed on windows
